Add some controllers,and exceptions

+++ src/main/java/com/victorku/musiccloud/service/AccountInfoServiceImpl.java
@@ -1,5 +1,6 @@
 package com.victorku.musiccloud.service;
 
+import com.victorku.musiccloud.exceptions.AccountIsNotExists;
 import com.victorku.musiccloud.model.AccountInfo;
 import com.victorku.musiccloud.repository.AccountInfoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +18,10 @@ public class AccountInfoServiceImpl implements AccountInfoService{
     }
 
     @Override
-    public void deleteAccountInfoById(Long id) {
+    public void deleteAccountInfoById(Long id) throws AccountIsNotExists {
+        if(!accountInfoRepository.exists(id)){
+            throw new AccountIsNotExists();
+        }
         accountInfoRepository.delete(id);
     }
